requirements_feseq.txt
torch>=1.10.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
pyyaml>=5.4.1
h5py>=3.6.0
pyarrow>=6.0.0
tqdm>=4.62.0

